compatibility fixes
