+++ src/components/RoomDropdown/MemberOptions.tsx
@@ -140,7 +140,7 @@ const UnstyledMemberOptions = ({ address }: any) => {
         address
     )
 
-    const { metamaskStreamrClient, roomId } = useStore()
+    const { metamaskStreamrClient, roomId, session: { streamrClient }  } = useStore()
 
     const revoke = useRevoker()
 
@@ -153,7 +153,8 @@ const UnstyledMemberOptions = ({ address }: any) => {
             stream,
         })
 
-        const members = await getRoomMembersFromStream(stream)
+        const members = await getRoomMembersFromStream(streamrClient!, stream.id)
+        console.info('members', members)
         dispatch({
             type: ActionType.SetRoomMembers,
             payload: {
