+++ src/hooks/useInviterSelf.ts
@@ -17,6 +17,8 @@ export default function useInviterSelf(): Inviter {
                 return
             }
 
+            console.info('calling inviteSelf for streams', streams.map(s => s.id))
+
             const tasks = streams.map((stream) => {
                 return {
                     streamId: stream.id,
